+++ app/processing/__init__.py
@@ -2,7 +2,8 @@ from .extract import extract_command, extract_all_command
 from .index_embeddings import index_embeddings_command, index_all_embeddings_command
 from .reduce import reduce_command, reduce_all_command
 from .index_audio import index_audio_command, index_all_audio_command
-from .metadata import load_jamendo_metadata_command
+from .metadata.jamendo import load_jamendo_metadata_command
+from .metadata.id3 import load_id3_metadata_command
 
 
 def init_app(app):
@@ -15,3 +16,4 @@ def init_app(app):
     app.cli.add_command(index_audio_command)
     app.cli.add_command(index_all_audio_command)
     app.cli.add_command(load_jamendo_metadata_command)
+    app.cli.add_command(load_id3_metadata_command)